refactor: extract GuardedExecutor + execution types to mandate-core (P5.1 prereq)
IP-4 prerequisite from docs/keeperhub-integration-paths.md: a future
mandate-keeperhub-adapter crate needs to depend on mandate-core alone,
not the whole workspace. Moving the trait + receipt + error types out of
mandate-execution into a new mandate_core::execution module unblocks
that — the IP-1 envelope helper (PR β / P5.1 proper) lands on top.

Changes:
- New crates/mandate-core/src/execution.rs hosting GuardedExecutor +
  ExecutionReceipt + ExecutionError. Imports come exclusively from
  mandate_core::aprp + mandate_core::receipt — no cross-crate use-site
  coupling, confirming Dev B's KH-audit "zero leaks at function-signature
  level" finding.
- crates/mandate-execution/src/lib.rs re-exports the moved symbols so
  every existing consumer (mandate-server, mandate-cli, mandate-mcp,
  demo-agents/research-agent) keeps compiling without an import-path
  change.
- keeperhub.rs / uniswap.rs internal `use crate::{...}` → `use
  mandate_core::execution::{...}` so the adapter modules consume the
  extracted trait directly.

Pure refactor — zero behaviour change. Tests stay 292/292; both demo
runners untouched (13-gate final demo green, production-shaped runner
26/0/1).
