+++ crates/mandate-execution/src/lib.rs
@@ -1,51 +1,31 @@
 //! Mandate execution: sponsor execution adapters.
 //!
-//! This crate exposes the `GuardedExecutor` trait that every sponsor adapter
-//! must implement. The contract is: an executor takes a Mandate-approved
-//! `PolicyReceipt` plus the original `PaymentRequest` and returns an
-//! `ExecutionReceipt` that callers can attach to the Mandate audit log.
+//! This crate exposes the [`GuardedExecutor`] trait that every sponsor
+//! adapter must implement. The contract is: an executor takes a
+//! Mandate-approved [`PolicyReceipt`](mandate_core::receipt::PolicyReceipt)
+//! plus the original [`PaymentRequest`](mandate_core::aprp::PaymentRequest)
+//! and returns an [`ExecutionReceipt`] that callers can attach to the
+//! Mandate audit log.
 //!
 //! * **Mandate decides.** The receipt is the proof of authorisation.
 //! * **Sponsor executes.** Each adapter is a thin wrapper over the partner's
 //!   real interface (or a clearly-disclosed local mock when credentials are
 //!   not available during the hackathon build).
+//!
+//! The trait + receipt + error types live in
+//! [`mandate_core::execution`] (extracted in PR #48 / P5.1 prereq) so the
+//! IP-4 standalone-adapter crate can depend on `mandate-core` alone. They
+//! are re-exported here so existing call sites
+//! (`mandate-server`, `mandate-cli`, `mandate-mcp`,
+//! `demo-agents/research-agent`) keep working without import-path
+//! changes.
 
 pub mod keeperhub;
 pub mod uniswap;
 
 pub use keeperhub::{KeeperHubExecutor, KeeperHubMode};
+pub use mandate_core::execution::{ExecutionError, ExecutionReceipt, GuardedExecutor};
 pub use uniswap::{
     evaluate_swap, SwapCheck, SwapPolicy, SwapPolicyOutcome, SwapQuote, SwapToken, UniswapExecutor,
     UniswapMode,
 };
-
-use mandate_core::aprp::PaymentRequest;
-use mandate_core::receipt::PolicyReceipt;
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum ExecutionError {
-    #[error("policy receipt rejected: decision={0:?}")]
-    NotApproved(mandate_core::receipt::Decision),
-    #[error("sponsor backend offline: {0}")]
-    BackendOffline(String),
-    #[error("integration: {0}")]
-    Integration(String),
-}
-
-#[derive(Debug, Clone)]
-pub struct ExecutionReceipt {
-    pub sponsor: &'static str,
-    pub execution_ref: String,
-    pub mock: bool,
-    pub note: String,
-}
-
-pub trait GuardedExecutor {
-    fn sponsor_id(&self) -> &'static str;
-    fn execute(
-        &self,
-        request: &PaymentRequest,
-        receipt: &PolicyReceipt,
-    ) -> Result<ExecutionReceipt, ExecutionError>;
-}
