+++ crates/mandate-execution/src/uniswap.rs
@@ -20,11 +20,10 @@
 use std::str::FromStr;
 
 use mandate_core::aprp::PaymentRequest;
+use mandate_core::execution::{ExecutionError, ExecutionReceipt, GuardedExecutor};
 use mandate_core::receipt::{Decision, PolicyReceipt};
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
-
-use crate::{ExecutionError, ExecutionReceipt, GuardedExecutor};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[serde(deny_unknown_fields)]
